fix: Cast u32 to usize for update_grid_stats calls
Fixed type mismatch errors on lines 207 and 275:
- grid_levels is u32 from config
- update_grid_stats expects usize
- Added safe type conversion

Compiles cleanly now!

+++ src/bots/grid_bot.rs
@@ -202,7 +202,8 @@ impl GridBot {
             self.grid_initialized = true;
             info!("✅ Initial grid placed successfully");
             
-            let total_levels = self.config.trading.grid_levels;
+            // 🔧 FIX: Cast u32 to usize
+            let total_levels = self.config.trading.grid_levels as usize;
             let used_levels = self.grid_state.count().await;
             self.enhanced_metrics.update_grid_stats(total_levels, used_levels);
             
@@ -270,7 +271,8 @@ impl GridBot {
         self.grid_repositions += 1;
         self.last_reposition_time = Some(std::time::Instant::now());
 
-        let total_levels = self.config.trading.grid_levels;
+        // 🔧 FIX: Cast u32 to usize
+        let total_levels = self.config.trading.grid_levels as usize;
         let used_levels = self.grid_state.count().await;
         self.enhanced_metrics.update_grid_stats(total_levels, used_levels);
